fix bug qua 2 ti khong cap nhat duoc

+++ src/main/java/com/technology/apigateway/controller/request/MainRequest.java
@@ -67,7 +67,7 @@ public class MainRequest {
     private String registrationNumber;
     private String engineNumber;
     private String chassisNumber;
-    private Integer vehiclePrice;
+    private Long vehiclePrice;
     private Integer loanAmount;
     private Integer term;
     private Integer saleId;
